fix: simplify ui a bit
